# frontend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/candidate/register": {
      "post": {
        "description": "Registers a new user and starts email verification flow. May return auth state depending on verification status.",
        "operationId": "AuthCandidateController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create new user account",
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/verify-email": {
      "post": {
        "description": "Validates email verification code and activates user account.",
        "operationId": "AuthCandidateController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true
                  }
                }
              }
            }
          }
        },
        "summary": "Verify email address",
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/login": {
      "post": {
        "description": "Validates credentials and returns auth state (tokens, MFA, onboarding, or verification).",
        "operationId": "AuthCandidateController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDtoSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Authenticate user",
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/logout": {
      "post": {
        "description": "Invalidates refresh token and clears session cookies.",
        "operationId": "AuthCandidateController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout user",
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
...
    "/auth/candidate/onboarding": {
      "post": {
        "description": "Finalizes onboarding using temporary auth token.",
        "operationId": "AuthCandidateController_completeOnboarding",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Complete onboarding",
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/github/link": {
      "get": {
        "operationId": "AuthCandidateController_linkGithub",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Link GitHub account",
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/github/link/callback": {
      "get": {
        "operationId": "AuthCandidateController_linkGithubCallback",
        "parameters": [
          {
            "name": "state",
            "required": true,
            "in": "query",
            "description": "Security state parameter for OAuth verification",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GitHub linking callback",
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/google/link": {
      "get": {
        "operationId": "AuthCandidateController_linkGoogle",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Link Google account",
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/google/link/callback": {
      "get": {
        "operationId": "AuthCandidateController_linkGoogleCallback",
        "parameters": [
          {
            "name": "state",
            "required": true,
            "in": "query",
            "description": "Security state parameter for OAuth verification",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Google linking callback",
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/refresh": {
      "post": {
        "operationId": "AuthCandidateController_refresh",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refresh tokens",
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/password-reset/request": {
      "post": {
        "description": "Starts the password reset flow. Always returns a generic success response.",
        "operationId": "AuthCandidateController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Request password reset",
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/password-reset/confirm": {
      "post": {
        "description": "Resets the password with a valid password reset token.",
        "operationId": "AuthCandidateController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Confirm password reset",
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/mfa/setup": {
      "get": {
        "operationId": "AuthCandidateController_setupMfa",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/mfa/activate": {
      "post": {
        "operationId": "AuthCandidateController_activateMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateMfaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/mfa/verify": {
      "post": {
        "operationId": "AuthCandidateController_verifyMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMfaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
    "/auth/candidate/mfa/verify-recovery": {
      "post": {
        "operationId": "AuthCandidateController_verifyMfaRecovery",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMfaRecoveryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth (candidates)"
        ]
      }
    },
...
  },
  "info": {
    "title": "a16zero Employer API",
    "description": "Backend MVP for Employer platform features and Account Abstraction auth verification.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "internal-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-key",
        "description": "Internal API key"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "description": "Valid email address",
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "username": {
            "description": "Unique username",
            "type": "string",
            "minLength": 3
          },
          "password": {
            "minLength": 8,
            "allOf": [
              {
                "pattern": "[A-Z]"
              },
              {
                "pattern": "[0-9]"
              },
              {
                "pattern": "[^A-Za-z0-9]"
              }
            ],
            "description": "Secure password with complexity requirements"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "CANDIDATE",
              "RECRUITER"
            ],
            "description": "User role in the system"
          }
        },
        "required": [
          "password",
          "role"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "6-digit verification code sent to the users email"
          }
        },
        "required": [
          "code"
        ]
      },
      "LoginDtoSchema": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "minLength": 1,
            "description": "Email or username"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "description": "Secret password"
          }
        },
        "required": [
          "identifier",
          "password"
        ]
      },
      "OnboardingDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 20,
            "description": "Unique public username"
          }
        },
        "required": [
          "username"
        ]
      },
      "RequestPasswordResetDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "description": "Account email address"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Password reset token received via email"
          },
          "newPassword": {
            "minLength": 8,
            "allOf": [
              {
                "pattern": "[A-Z]"
              },
              {
                "pattern": "[0-9]"
              },
              {
                "pattern": "[^A-Za-z0-9]"
              }
            ],
            "description": "New secure password"
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "ActivateMfaDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6,
            "description": "6-digit MFA code"
          }
        },
        "required": [
          "code"
        ]
      },
      "VerifyMfaDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "Internal User ID"
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6,
            "description": "6-digit MFA code"
          },
          "mfaToken": {
            "type": "string",
            "minLength": 1,
            "description": "Temporary MFA session token"
          }
        },
        "required": [
          "userId",
          "code",
          "mfaToken"
        ]
      },
      "VerifyMfaRecoveryDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "Internal User ID"
          },
          "backupCode": {
            "type": "string",
            "minLength": 8,
            "maxLength": 8,
            "description": "8-character recovery code"
          },
          "mfaToken": {
            "type": "string",
            "minLength": 1,
            "description": "Temporary MFA session token"
          }
        },
        "required": [
          "userId",
          "backupCode",
          "mfaToken"
        ]
      },
...
    }
  }
}